chore: zero-warning pass; escalation test docs match actual behavior
- Rename reserve_escalation_beats_two_lying_workers ->
  reserve_escalation_beats_lying_worker: round 1 is one liar + one
  honest worker, so the old name/doc overstated the scenario. With a
  pool of 3, two round-1 liars can never end in Accept (decide()
  requires a strict majority), so the code was right and the docs
  were wrong. DESIGN.md's 'two lying with DIFFERENT fabricated
  results' description corrected the same way.
- Drop the stale flake NOTE from the test: the root cause (dispatch
  firing before full-pool auth) was fixed in 48ba668; comment now
  states the regression-guard purpose.
- Clippy clean under 1.98.1 lints: remove dead imports/locals
  (TcpStream, ChunkEnd, session_loop peer_ip, daemon listen_port
  shadow), no-op casts, needless range loops, single-arm match ->
  let, QueuedJob type alias. Hot fetch/alignment checks keep '%'
  (is_multiple_of is an out-of-line call in debug builds) with
  file-level allows documenting why.

+++ crates/rvcore/tests/isa.rs
@@ -350,7 +350,7 @@ fn input_output_abi_roundtrip() {
     let output = out.output.expect("output present");
     // The program stored the input length to the OUTPUT_LEN header,
     // which the emulator consumed; the data region itself is zeros.
-    assert_eq!(output.len(), input.len() as usize);
+    assert_eq!(output.len(), input.len());
     assert!(output.iter().all(|&b| b == 0));
     assert_eq!(mem.read(INPUT_LEN_ADDR, 8).unwrap(), input.len() as u64);
     assert_eq!(mem.read(INPUT_DATA_ADDR, 1).unwrap(), b'h' as u64);
